same approach for the regular input
